+++ NEWS.md
@@ -1,10 +1,12 @@
 Changelog
 
 # version 0.4.1
+16th September 2020
 
 ## Changes
 
 * add 10-year cardiovascular (CVD) risk `frisk10` to QIM 08.
+* increase maximum upload filesize to 300 megabytes
 
 # version 0.4.0
 18th August 2020